feat(jdk26): add demo for JEP 526
