+++ src/main/java/org/javademos/init/Java26DemoLoader.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import org.javademos.commons.IDemo;
 import org.javademos.commons.IDemoLoader;
 
-// import org.javademos.java26.jepXXX.YYYDemo;
+import org.javademos.java26.jep526.LazyConstantsDemo;
 
 /**
  * Loads demos for Java 26.
@@ -14,6 +14,6 @@ public class Java26DemoLoader implements IDemoLoader {
     
     @Override
     public void loadDemos(Map<Integer, IDemo> demos) {
-        // demos.put(XXX, new YYYDemo());
+        demos.put(526, new LazyConstantsDemo());
     }
 }
