+++ src/main/resources/JDK25Info.json
@@ -11,9 +11,9 @@
     "jep": 502,
     "jdk": 25,
     "name": "JEP 502 - Stable Values (Preview)",
-    "dscr": "Stable Values offer an immutable data abstraction for performance and safety.",
-    "link": false,
-    "code": true
+    "dscr": "Up-to-date demo in `org.javademos.java26.jep526.LazyConstantsDemo`.",
+    "link": true,
+    "code": false
   },
   {
     "jep": 503,
