+++ src/main/resources/JDK26Info.json
@@ -1,2 +1,10 @@
-[
+[  
+  {
+    "jep": 526,
+    "jdk": 26,
+    "name": "JEP 526 - Lazy Constants (Preview)",
+    "dscr": "Lazy Constants offer an immutable data abstraction for performance and safety.",
+    "link": false,
+    "code": true
+  }
 ]
